requirements.txt
# Work Rules Updater - Dependencies

# CLI
click>=8.1.0
rich>=13.0.0

# 법령 API
requests>=2.31.0
lxml>=5.0.0

# 문서 처리
python-docx>=1.1.0

# 환경변수
python-dotenv>=1.0.0

# 데이터 처리
pydantic>=2.0.0
